fix interact text rendered on top of skill menu
In some situations the text's hitbox would partially block the skill add/remove buttons. This has been fixed by changing the sort order of the mod's UI canvas to a value higher than the game canvas.

This is a quick fix - a proper fix will be added in the major update.