partial rework of xp tables
A new solution for xp tables has been implemented. It is now much easier to override the embedded xp tables and add custom xp tables - all without having to worry about messing up.

The embedded xp tables will always load. Custom xp tables will be skipped and log a warning if they fail to load, allowing the mod to continue working because of the embedded xp tables.

This significantly improves the UX - meaning it is less error-prone and more stable.