rework and cleanup A work in progress rework and cleanup. This commit is big and unstable - it will break the majority of the mod.
The majority of the files have only been modified to clean up the code style.

The files with actual changes are the ones related to leveling. I'm working on reducing calls to the server by making the ZNetView owners responsible for handling the calculations. I will implement syncing for configs so that the clients calculate based on the server's configs.