Make it easier create max experience formulas
The config option to select a formula is not being updated dynamically yet.

The plan is to allow other mods to create custom formulas. I also need to implement a way for formulas to access more parameters.

Maybe change the EvaluateMaxExperience interface method to accept an interface instead. That way it should be straightforward to supply it with a struct that contains all the data it needs.